fix(vim): `xn` short after together jumped to next search
I had a very annoying bug where, shortly after pressing some command and
then trying to move cursor down (in my map "mnei" it was "n"), which for
some reason jumped to the next search. It was very annoying and it looks
like I had some left over configuration that I was not using much (I
don't even know if it was working). Removing it seems to fix it.