feat(vim): Try centering the cursor all the time
I want to experiment with this. And after some very short time I think I
like it. The nice thing is that you don't have to search for the cursor
and it stays on the predictable place. It is much better to use with
cursorline.

The only issue I have with it for now is that my current document
highlight is the same color as the cursorline, which makes it that you
don't see that the symbol under the cursor is highlighted, but you see
all others. I will have to try to update the LspReferenceText highlight.