feat(vim): Disable syntax for no plugin vim
For some reason every time I open a file I get the default highlighting,
I need to do :e to reset it and only then I get the tree-sitter
highlighting. This is pretty annoying and I was not able to find similar
issue in tree-sitter repo so I am disabling it for now.