feat(vim): Unbind K in visual selection mode
Sometimes I am too fast when using visual line selection and press `K`
instead of `k`. Which tries to open a documentation for the selected
text. It takes a few moments for search to return that it was not able
to find anything. So just try to unbind it instead.