perf(git): remove submoduleSummary from git config
This makes the git status command faster. From my random measurements it
was 200ms before and only up to 50ms after this change.

When I am making commits, I usually run git status, and since it was
slow, it was eating my stdin (pretyped next command), which was very
annoying.